refactor: move series spec to siuba.op.support
